full client and server validations with zod and react hook form

+++ src/actions/actions.ts
@@ -3,15 +3,21 @@
 import prisma from "@/lib/db";
 import { PetInternal } from "@/lib/types";
 import { sleep } from "@/lib/utils";
+import { petFormSchema, petIdSchema } from "@/lib/validations";
 import { Pet } from "@prisma/client";
 import { revalidatePath } from "next/cache";
 
-export async function addPet(pet: PetInternal) {
+export async function addPet(pet: unknown) {
   await sleep(2500);
+
+  const validatedPet = petFormSchema.safeParse(pet);
+  if (!validatedPet.success) {
+    return { message: "Invalid pet data" };
+  }
 
   try {
     await prisma.pet.create({
-      data: pet,
+      data: validatedPet.data,
     });
   } catch (error) {
     return { message: "Could not add pet" };
@@ -20,15 +26,21 @@ export async function addPet(pet: PetInternal) {
   revalidatePath("/app", "layout");
 }
 
-export async function editPet(petId: Pet["id"], newPetData: PetInternal) {
+export async function editPet(petId: unknown, newPetData: unknown) {
   await sleep(2500);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+  const validatedPet = petFormSchema.safeParse(newPetData);
+  if (!validatedPet.success || !validatedPetId.success) {
+    return { message: "Invalid pet data" };
+  }
 
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: newPetData,
+      data: validatedPet.data,
     });
   } catch (error) {
     return { message: "Could not edit pet" };
@@ -37,13 +49,19 @@ export async function editPet(petId: Pet["id"], newPetData: PetInternal) {
   revalidatePath("/app", "layout");
 }
 
-export async function checkoutPet(petId: Pet["id"]) {
+export async function checkoutPet(petId: unknown) {
   await sleep(2500);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+
+  if (!validatedPetId.success) {
+    return { message: "Invalid pet data" };
+  }
 
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
   } catch (error) {
